requirements.txt
# Schema RAG Dependencies

# Core dependencies
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Optional: For SQLAlchemy model export (if using export script)
# sqlalchemy>=2.0.0
